+++ oscar-cbls/src/main/scala/oscar/cbls/lib/invariant/numeric/Advanced.scala
@@ -39,7 +39,7 @@ case class SumConstants(vars: Array[Long], cond: SetValue)
   registerStaticAndDynamicDependency(cond)
   finishInitialization()
 
-  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Long], removedValues: Iterable[Long], oldValue: SortedSet[Long], newValue: SortedSet[Long]): Unit = {
+  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Int], removedValues: Iterable[Int], oldValue: SortedSet[Int], newValue: SortedSet[Int]): Unit = {
     for (added <- addedValues)  this :+= vars(added)
     for (deleted <- removedValues) this :-= vars(deleted)
   }
@@ -91,13 +91,13 @@ case class SumElements(vars: Array[IntValue], cond: SetValue)
     this :+= (NewVal - OldVal)
   }
 
-  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Long], removedValues: Iterable[Long], oldValue: SortedSet[Long], newValue: SortedSet[Long]): Unit = {
+  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Int], removedValues: Iterable[Int], oldValue: SortedSet[Int], newValue: SortedSet[Int]): Unit = {
     for (added <- addedValues) notifyInsertOn(v: ChangingSetValue, added)
     for(deleted <- removedValues) notifyDeleteOn(v: ChangingSetValue, deleted)
   }
 
   @inline
-  def notifyInsertOn(v: ChangingSetValue, value: Long) {
+  def notifyInsertOn(v: ChangingSetValue, value: Int) {
     assert(v == cond)
     assert(keyForRemoval(value) == null)
     keyForRemoval(value) = registerDynamicDependency(vars(value),value)
@@ -106,7 +106,7 @@ case class SumElements(vars: Array[IntValue], cond: SetValue)
   }
 
   @inline
-  def notifyDeleteOn(v: ChangingSetValue, value: Long) {
+  def notifyDeleteOn(v: ChangingSetValue, value: Int) {
     assert(v == cond)
     assert(keyForRemoval(value) != null)
     keyForRemoval(value).performRemove()
@@ -147,17 +147,17 @@ case class ProdConstants(vars: Array[Long], cond: SetValue)
     }
   }
 
-  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Long], removedValues: Iterable[Long], oldValue: SortedSet[Long], newValue: SortedSet[Long]): Unit = {
+  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Int], removedValues: Iterable[Int], oldValue: SortedSet[Int], newValue: SortedSet[Int]): Unit = {
     for (added <- addedValues) notifyInsertOn(v: ChangingSetValue, added)
     for(deleted <- removedValues) notifyDeleteOn(v: ChangingSetValue, deleted)
   }
 
   @inline
-  def notifyInsertOn(v: ChangingSetValue, value: Long) {
+  def notifyInsertOn(v: ChangingSetValue, value: Int) {
     assert(v == cond)
 
     if(vars(value) == 0L){
-      NullVarCount += 1L
+      NullVarCount += 1
     }else{
       NonNullProd *= vars(value)
     }
@@ -165,10 +165,10 @@ case class ProdConstants(vars: Array[Long], cond: SetValue)
   }
 
   @inline
-  def notifyDeleteOn(v: ChangingSetValue, value: Long) {
+  def notifyDeleteOn(v: ChangingSetValue, value: Int) {
 
     if(vars(value) == 0L){
-      NullVarCount -= 1L
+      NullVarCount -= 1
     }else{
       NonNullProd = NonNullProd / vars(value)
     }
@@ -230,10 +230,10 @@ case class ProdElements(vars: Array[IntValue], cond: SetValue)
     assert(vars(index) == v)
     assert(keyForRemoval(index)!=null)
     if (OldVal == 0L && NewVal != 0L){
-      NullVarCount -=1L
+      NullVarCount -=1
       NonNullProd *=NewVal
     }else if(OldVal != 0L && NewVal == 0L){
-      NullVarCount +=1L
+      NullVarCount +=1
       NonNullProd =NonNullProd/OldVal
     }else{
       NonNullProd = NonNullProd/OldVal
@@ -242,19 +242,19 @@ case class ProdElements(vars: Array[IntValue], cond: SetValue)
     affectOutput()
   }
 
-  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Long], removedValues: Iterable[Long], oldValue: SortedSet[Long], newValue: SortedSet[Long]): Unit = {
+  override def notifySetChanges(v: ChangingSetValue, id: Int, addedValues: Iterable[Int], removedValues: Iterable[Int], oldValue: SortedSet[Int], newValue: SortedSet[Int]): Unit = {
     for (added <- addedValues) notifyInsertOn(v: ChangingSetValue, added)
     for(deleted <- removedValues) notifyDeleteOn(v: ChangingSetValue, deleted)
   }
 
   @inline
-  def notifyInsertOn(v: ChangingSetValue, value: Long) {
+  def notifyInsertOn(v: ChangingSetValue, value: Int) {
     assert(v == cond)
     assert(keyForRemoval(value) == null)
     keyForRemoval(value) = registerDynamicDependency(vars(value),value)
 
     if(vars(value).value == 0L){
-      NullVarCount += 1L
+      NullVarCount += 1
     }else{
       NonNullProd *= vars(value).value
     }
@@ -262,7 +262,7 @@ case class ProdElements(vars: Array[IntValue], cond: SetValue)
   }
 
   @inline
-  def notifyDeleteOn(v: ChangingSetValue, value: Long) {
+  def notifyDeleteOn(v: ChangingSetValue, value: Int) {
     assert(v == cond)
     assert(keyForRemoval(value) != null)
 
@@ -270,7 +270,7 @@ case class ProdElements(vars: Array[IntValue], cond: SetValue)
     keyForRemoval(value) = null
 
     if(vars(value).value == 0L){
-      NullVarCount -= 1L
+      NullVarCount -= 1
     }else{
       NonNullProd = NonNullProd / vars(value).value
     }
